refactor: group toast in a dir and lift CopyText into base
