+++ apps/web/src/base/Toast/index.ts
@@ -1,0 +1,2 @@
+export { default as Toaster } from "./Toaster";
+export { toast } from "./useToast";
