+++ apps/web/src/routes/_authenticated.tsx
@@ -5,7 +5,7 @@ import * as Sse from "@app/sse/SseConnection";
 import Banner from "@banner/components/Banner";
 import LoadingPlaceholder from "@base/LoadingPlaceholder";
 import SkipLink from "@base/SkipLink";
-import Toaster from "@base/Toaster";
+import { Toaster } from "@base/Toast";
 import Nav from "@nav/components/Nav";
 import Sidebar from "@nav/components/Sidebar";
 import * as Sentry from "@sentry/tanstackstart-react";
